Added support for restart_policy parameter
This patch adds the restart_policy parameter to services. The parameter
is passed to the Docker API when starting containers. Its recognized values
are the same as in the Docker CLI ("always" and "on-failure[:max-retry]").

For future compatibility, the policy can also be specified as a
dictionary, which is passed directly to the Docker API (after converting
the keys to CamelCase).

This feature seems to have been requested at least in issue #478.

Signed-off-by: Kenneth Falck <kennu@iki.fi>