docker/fig#761 fig.yml to specify a template to override a service.
Specifying a key of "template" within a service can refer to another yml
file causing the service to "override" another service in the template
file.

The value of template can either be a string, which refers to a service
of the same name in another yml file, or a dict consisting of two keys:
"service" and "path".

Signed-off-by: Gerald Kaszuba <gak@gak0.com>