Add commit to run command
Added a commit option on the run command, this will help to improve
productivity with various package manages. For example if Im building a
rails app and I would like to change/update a dependancy. I would have
to rebuild the docker image every time which can take some time. If
after the run command it could commit the changes made to the docker
image then it would save a load of time.

Signed-off-by: Richard Adams <richard@madwire.co.uk>