Expand Expose statement to include IPAddr access
Since I only use Docker Containers directly via their IP address I don't bind any ports of my containers directly to my Host Machine. If the expose for Docker Compose works just like that of the Dockerfile then this definition should be updated to reflect that a port can be exposed on the container and then accessible by using the IP address of the container along with the port specified in the "EXPOSE' statement.