Fix NoneType on smart recreate
This fixes issue #1561

smart-recreate hashes a dictionary of properties of the service, one of
those is the image. The image can be null if it has not yet been
fetched. By leaving it out of the hash in this case it will always
mismatch causing the change to be recognised.

Signed-off-by: Adam Scarr <adam@vektah.net>