Handle windows volume paths
When a relative path is expanded and we're on a windows platform,
it expands to include the drive, eg C:\ , which was causing a ConfigError
as we split on ":" in parse_volume_spec and that was giving too many parts.
Re-joining the first two parts if we're on a windows platform allows the
rest of the logic for parsing volume spec to work.

Also moved IS_WINDOWS_PLATFORM into a const to make it accessible rather than
writing a check each time for sys.platform.

Signed-off-by: Mazz Mosley <mazz@houseofmnowster.com>