Enable interpolation of keys in compose file
This adds interpolation of all keys in the compose file except the
sections themselves. I have a need for doing that when, for example,
creating overlay networks based on some unique name:

```
networks:
  ${NAMESPACE}_overlay:
     driver: overlay

services:
  my_service:
    networks:
      - "${NAMESPACE}_overlay"
```

The reason for the slight code restructuring was to avoid code
duplication when implementing this. I do welcome feedback.

Signed-off-by: Mattias Jansson <mattias.jansson@erlang-solutions.com>