Replace PyYAML with ruamel.yaml
PyYAML supports YAML 1.1 which has features that require users
to quote some input. Main problem are sexagesimals, so that
"- 25:25" is interpreted as "- 1525". In ruamel.yaml this is
intpreted as if it was written "- '25:25'".
In addition the above, Yes/No/On/Off are no longer interpreted
as booleans, and scalars starting with a 0 and consisting of
numbers no longer as octals, but normal integers (use 0o12 if you
need octals).

The YAML 1.2 is implemented in ruamel.yaml for the RoundTripLoader.
In order to preserver safe loader, the VersionedResolver is combined
with the SafeConstructor in SafeLoader_1_2, which is used explicitly
to load YAML files.
A simple test is provided to confirm that sexagesimals no longer
break user input.

(Several lines that I never touched needed to be wrapped to let the
pre-commit hooks pass the flake8 test).

Signed-off-by: Anthon van der Neut <a.van.der.neut@ruamel.eu>