Add first attempt at supporting test suite on ARM
Modify `script/test/default` so it supports a first attempt at testing on ARM.
Call the script with: `$ DOCKERFILE=Dockerfile.armhf script/test/default` to
use the Dockerfile.armhf instead of the default Dockerfile for building the image.

However, running the script is not working fully. The problem is that
`dockerswarm/dind` does not provide an ARM image and therefore cannot be executed.
If that is fixed then we should be able to change the script in order to use the ARM
image instead of the default x86_64 image for running further tests.

Signed-off-by: Jean-Christophe Berthon <huygens@berthon.eu>