Replace "_" with "-" in build_container_name
Replace "_" with "-" in build_container_name to create RFC 952 (https://tools.ietf.org/html/rfc952) compliant hostnames.
The valid tokens as per RFC are A-Z (case insensitive),0-9 and - (hyphen) with the first character being alphabetic.
RFC 1123 (https://tools.ietf.org/html/rfc1123) expands on this allowing the first character in the label to be a digit too.

Signed-off-by: Terror <miguel.a.terron@gmail.com>