Pass the --disable-pip-version-check option to pip freeze
We invoke pip as part of a check for an incompatible package and by
default, pip performs a version check to see whether it is up to date.

This check is unnecessary for our us - we just want a list of currently
installed packages. Prevent this check by supplying the
'--disable-pip-version-check' flag.

Signed-off-by: Charles Bailey <cbailey32@bloomberg.net>