Capture the beginning of service output during run
Previously the container was started before pseudoterminal was started,
therefore some of the output could've been lost. Goal of this commit is
to make sure that all the input is printed out. To that end, the run
command executes the following sequence:

1. connect container to networks
2. start pty
3. start container

The original code was introduced in
f3e55568d1a7c111398876bec84da8ed8b42da7f

Signed-off-by: Jarek Lipski <loomchild@loomchild.net>