Show what's in progress when running parallel
Prior to this commit, docker-compose only showed "done" or "error" once
a parallel task was finished. This made it hard to tell which task(s) were
actually in progress.

Now there's an additional "working" status that's shown in yellow when a task
is in-progress.

I implemented this by adding an additional "status" field to the result tuple
in parallel.py.

Updates #4556

Signed-off-by: Evan Shaw <evan@vendhq.com>