Allow using COMPOSE_PROJECT_NAME in environment substitutions
When the `COMPOSE_PROJECT_NAME` is defined in the calling environment, it is
available for substitution in the `docker-compose.yml` file.  However, this
cannot be relied upon because the project name can be inferred from the
parent directory name.

This change ensures the `COMPOSE_PROJECT_NAME` environment variable is always
available for subtitution in `docker-compose.yml` file regardless of the
provenance of the project name.

Signed-off-by: Andre Caron <andre.l.caron@gmail.com>