Add abstraction for buildkit CLI args
Rather than putting buildkit strings directly in compose file, use a mapping abstraction. The format is slightly odd looking string or dict as the only truly required mount arg is id but we may also want to provide a whole dict of options including source and any other flags that buildkit supports for a mount spec.

Signed-off-by: Ollie Edwards <oliver.s.edwards@gmail.com>