Speed up fish completions for subcommands
The current method of generating completions for the fish shell relies
on running `docker-compose --help`. On my system (a 2018 MacBook Pro)
this routinely takes several hundred milliseconds:

```
$ hyperfine --warmup 3 'docker-compose --help'
Benchmark #1: docker-compose --help
  Time (mean ± σ):     351.6 ms ±  16.0 ms    [User: 295.5 ms, System: 44.3 ms]
  Range (min … max):   332.8 ms … 384.0 ms    10 runs
```

If I configure fish to automatically load these completions, it makes
launching a new fish process very sluggish.

In addition, the regular expression used to parse the output of
`docker-compose --help` incorrectly generates completions for `name` and
`in` due to the following lines:

https://github.com/docker/compose/blob/9c5351cf27d10a7b8078496fb21a37929b6687cb/compose/cli/main.py#L208
https://github.com/docker/compose/blob/9c5351cf27d10a7b8078496fb21a37929b6687cb/compose/cli/main.py#L212

This commit fixes these issues by explicitly listing subcommands instead
of generating them dynamically. This means that adding/removing
subcommands won't automatically update the completions but according to
the git blame for `compose/cli/main.py`, no one has touched the list of
subcommands in the help message in 3+ years so I hope that won't be an
issue.

Signed-off-by: Collin Styles <collingstyles@gmail.com>