Add auto_up to exclude service from `docker-compose up`
This adds the new config option auto_up to services. When this is
explicitly set to false the service will not be started when executing
`docker-compose up` without service names (unless it is pulled in as a
dependency of another service).
If omitted auto_up defaults to true, i.e. start the service by default
when docker-compose is called without explicit service names.

This allows to define "optional" services in a docker-compose.yml which
are only started when listed explicitly in the command line like
docker-compose up optional_service

This closes #6742, #1896 and many more reports requesting this feature.
See also #3047

Signed-off-by: Roman Anasal <roman.anasal@academyconsult.de>