Support use of shell commands at env variable
This adds support for the following syntax: ${var:~(command)}, ${var:~command}, ${var~(command)} and ${var~command}
Allowing the use a shell command to assign a value to a variable when the var is not already set at the host.

For example: "HUID_ENV: $(HUID_HOST:~(id -u))" would assign the value returned by "id -u" (UID of the current user)
to the environment variable "HUID_ENV", unless "HUID_HOST" was already set and not empty.

Signed-off-by: Gustavo Moura <gusdev77@gmail.com>