run docker-compose with remote's local socket
This commit will use /var/run/docker.sock if it appears to be a socket within the docker container.
For local runs this would not add benefits but for remote runs of docker, this would allow using a docker host that is only known to the host that is running this shell script.

Also:
ignore some errors,
use long-options,
keep debugging commands as comments.

Signed-off-by: immeëmosol <will+developer@willfris.nl>